bin/cellpose_segment.py
#!/usr/bin/env python3
"""Cellpose-based cell/spot segmentation on a TIFF image."""

import argparse, json, warnings
from pathlib import Path
import numpy as np
import matplotlib; matplotlib.use('Agg')
import matplotlib.pyplot as plt
warnings.filterwarnings('ignore')

def parse_args():
    p = argparse.ArgumentParser()
    p.add_argument('--image',       required=True)
    p.add_argument('--out_mask',    required=True)
    p.add_argument('--out_overlay', required=True)
    p.add_argument('--out_bounds',  required=True)
    p.add_argument('--out_stats',   required=True)
    return p.parse_args()

def main():
    args = parse_args()
    from PIL import Image
    import tifffile

    img = np.array(Image.open(args.image).convert('RGB'))

    try:
        from cellpose import models
        model = models.Cellpose(gpu=False, model_type='cyto')
        masks, flows, styles, diams = model.eval(img, diameter=None, channels=[0, 0])
    except Exception:
        # Fallback: simple threshold segmentation
        gray = img.mean(axis=2)
        from scipy import ndimage
        thresh = gray > gray.mean()
        labeled, n_cells = ndimage.label(thresh)
        masks = labeled.astype(np.int32)
        n_cells_found = n_cells
    else:
        n_cells_found = int(masks.max())

    # Save mask
    tifffile.imwrite(args.out_mask, masks.astype(np.int32))

    # Overlay
    fig, axes = plt.subplots(1, 2, figsize=(12, 6))
    axes[0].imshow(img); axes[0].set_title('Original')
    axes[1].imshow(masks, cmap='nipy_spectral'); axes[1].set_title(f'Segmentation ({n_cells_found} cells)')
    plt.tight_layout()
    plt.savefig(args.out_overlay, dpi=150); plt.close()

    # Cell boundaries CSV
    import pandas as pd
    from scipy import ndimage
    rows = []
    for cid in range(1, min(n_cells_found + 1, 5001)):
        yx = np.argwhere(masks == cid)
        if len(yx):
            cy, cx = yx.mean(axis=0)
            rows.append({'cell_id': cid, 'centroid_y': cy, 'centroid_x': cx,
                         'area': len(yx)})
    pd.DataFrame(rows).to_csv(args.out_bounds, index=False)

    stats = dict(n_cells=n_cells_found, image_shape=list(img.shape))
    Path(args.out_stats).write_text(json.dumps(stats, indent=2))
    print(f"[cellpose] Segmented {n_cells_found} cells")

if __name__ == '__main__':
    main()
