requirements.txt
# Core scientific computing
scanpy>=1.9.6
anndata>=0.9.2
squidpy>=1.3.0
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Spatial / image
cellpose>=2.2.0
tifffile>=2023.1.23
Pillow>=9.4.0

# Deep learning (GNN)
torch>=2.0.0
torch-geometric>=2.3.0

# Cell-cell communication (optional — pipeline uses Python re-implementations)
# liana>=1.0.0

# Reporting
jinja2>=3.1.0
h5py>=3.8.0
